v1
This code defines three structs, InterestRate, Balance, and Trade, which represent the data returned by the blockchain. The structs use the serde crate to support serialization and deserialization, which will make it easy to convert the data returned by the blockchain into these structs for use in your code.

The InterestRate struct has two fields: rate and asset. The rate field is a floating-point number representing the interest rate for the specified asset.

The Balance struct also has two fields: balance and asset. The balance field is a floating-point number representing the balance of the specified asset.

The Trade struct has five fields: buy_amount, sell_amount, buy_asset, sell_asset and timestamp. The buy_amount and sell_amount fields are floating-point numbers representing the amount of assets to buy and sell respectively. The buy_asset and sell_asset fields are strings representing the assets being bought and sold. The timestamp field is an int64 representing the timestamp of the trade.