v1
Note that: The code above is an example of a Cargo.toml file for a rust package. The package name is "my_trading_bot", the version is "0.1.0", and the author is "Your Name your@email.com". It lists the dependencies, such as "web3", "ethcontract", "ethkey" and "structopt".
Each of these dependencies will be used by your project in different ways. For example, web3 is a library for interacting with Ethereum nodes, ethcontract is a library for creating and interacting with smart contracts, ethkey is a library for managing Ethereum private keys, and structopt is a library for parsing command line arguments.
You should check the latest version of the dependencies when you start the project